Implementing Code in Promote Method
Here we add code to the promote method and in the Main method we attempt to promote the customer.  In the Promote method however we reinitialize the List of Orders.  Lets see what hapens.

Note:  When you run the code you are allowed to rese the Customers LIst of Orders this essentially empties the customers shopping cart.