Cleaning Up the Promote Method
Removed the buggy line of code in the promote method and add comment in the code to remind you to add code later to this method.

Note:  In this project the content of the Promote mehtod is not important.  It is the fact that by using the readonly modifier on a field will force the List to be initialized upon implementation of the class object itself.